requirements.txt
openai
yfinance
pandas
datetime
altair
matplotlib
seaborn
streamlit-chat
statsmodels
prophet
pandas_datareader
newspaper3k
konlpy
getmac
msoffcrypto-tool
openpyxl
pygwalker
langchain
tiktoken
chromadb
pykrx